target selection module deleted, relevant logic moved to linker. Closes #43
