+++ soweego/linker/resources/discogs_properties_mapping.json
