+++ soweego/linker/resources/musicbrainz_url_formatters.json
@@ -1,0 +1,25 @@
+{
+  "?allmusiccom": "https://www.allmusic.com/artist/$1",
+  "?rateyourmusiccom": "https://rateyourmusic.com/artist/$1",
+  "?lastfm": "https://www.last.fm/music/$1",
+  "?discogscom": "https://www.discogs.com/artist/$1",
+  "?imdbcom": "https://tools.wmflabs.org/wikidata-externalid-url/?p=345&url_prefix=https://www.imdb.com/&id=$1",
+  "?facebookcom": "https://www.facebook.com/$1",
+  "?openspotifycom": "https://open.spotify.com/artist/$1",
+  "?twittercom": "https://twitter.com/$1",
+  "?vgmdbnet": "https://vgmdb.net/artist/$1",
+  "?myspacecom": "https://myspace.com/$1",
+  "?viaforg": "https://viaf.org/viaf/$1",
+  "?dnbinfo": "https://d-nb.info/gnd/$1",
+  "?metalarchivescom": "http://www.metal-archives.com/artist.php?id=$1",
+  "?itunesapplecom": "https://itunes.apple.com/ca/artist/$1",
+  "?soundcloudcom": "https://soundcloud.com/$1",
+  "?youtubecom": "https://www.youtube.com/channel/$1",
+  "?imslporg": "https://imslp.org/wiki/$1",
+  "?instagramcom": "https://www.instagram.com/$1/",
+  "?secondhandsongscom": "https://secondhandsongs.com/artist/$1",
+  "?deezercom": "http://www.deezer.com/artist/$1",
+  "?playgooglecom": "https://play.google.com/store/music/artist?id=$1",
+  "?vkcom": "https://vk.com/$1",
+  "?bandcampcom": "https://$1.bandcamp.com"
+}
