--- soweego/target_selection/__init__.py
@@ -1,1 +1,0 @@
-from soweego.target_selection import bne, cli, commons, musicbrainz
