--- soweego/target_selection/bne/__init__.py
@@ -1,1 +1,0 @@
-import soweego.target_selection.bne.bne_baseline_matcher
