--- soweego/target_selection/commons/__init__.py
@@ -1,1 +1,0 @@
-from soweego.target_selection.commons import cli, matching_strategies
