--- soweego/target_selection/commons/cli.py
@@ -1,25 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""TODO module docstring"""
-
-__author__ = 'Marco Fossati'
-__email__ = 'fossati@spaziodati.eu'
-__version__ = '1.0'
-__license__ = 'GPL-3.0'
-__copyright__ = 'Copyleft 2018, Hjfocs'
-
-import click
-
-from soweego.target_selection.commons import matching_strategies
-
-CLI_COMMANDS = {
-    'baseline_match': matching_strategies.baseline
-}
-
-
-@click.group(name='commons', commands=CLI_COMMANDS)
-@click.pass_context
-def cli(ctx):
-    """Shared operations."""
-    pass
