--- soweego/target_selection/commons/constants.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""target_selection constants"""
-
-# path
-OUTPUT_LOCATION = 'output'
-RESOURCES_LOCATION = 'resources'
-TARGHET_SELECTION_LOCATION = 'soweego.target_selection'
-WIKIDATA_LOCATION = 'soweego.wikidata'
-
-WIKIDATA_SAMPLES = 'wikidata_sample.json'
