--- soweego/target_selection/discogs/__init__.py
@@ -1,1 +1,0 @@
-from soweego.target_selection.discogs import cli, baseline_matcher
