--- soweego/target_selection/musicbrainz/__init__.py
@@ -1,2 +1,0 @@
-from soweego.target_selection.musicbrainz import (cli,
-                                                  musicbrainz_baseline_matcher)
